Fix colors in host memory chart
The memory chart for a host has a few problems with the colors. The
colors change every time a particular memory type is displayed (or
hidden) and there is no legend (or another way to correlate the memory
type with the lines). Fix this by associating a color to each 'tag' and
updating the colors used for each series any time a series is added or
removed.

The list of chart colors was taken from:
http://chandoo.org/wp/2010/04/23/chart-colors-poll/

Reviewed-by: rkennke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-July/002309.html
PR1083