Add system.bundle fragment
This adds a dummy fragment which attaches to the system.bundle
(Bundle-SymbolicName). It's sole purpose is to have JVM internal packages
exported in an OSGi sense. Conceptually this is very similar to
osgi-export.properties.

I've also added poms for building this fragment. Building other eclipse
plug-ins won't be different. This fragment can be built using Tycho via:

$ cd client/eclipse
$ mvn clean install

I've not added it (yet) as a module to the main pom so that it gets built
together with the main thermostat build. This is because it doesn't do much and
is pretty useless on it's own. Once the Eclipse client prototype is in enabling
builds that way will make more sense.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-July/002525.html