Make search work for hosts
When deciding to show a host or VM in the tree, we require one of two
possible conditions to succeed: either the node matches all the filters,
or a child node needs to be displayed. For a search term which just
matches the host name, all the filters need to be matched. There are
currently two filters: the search text filter and the LivingVmFilter.
The LivingVMFilter does not match hosts, of course. The result is that
if no VMs are selected by the search query, the host is not shown
either.

Fix this by using separate filters for different parts of the tree:
HostFilter and VmFilter (and matching HostDecorator and VmDecorator).
The searchFilter is now used as both a host and vm filter, but other
filters are for host or vms only.

Reviewed-by: jerboaa, neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-July/002532.html