Fix CPU Chart
Instead of the average machine load, show the utilization of each
processor based on informatio from /proc/stat.

Also adds support for storing arrays in storage. It was needed for
storing the information of a variable number of CPUs which varies
depending on the machine.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-August/002637.html