Add target definition file for Fedora Eclipse
This patch adds a target definition file in order for new developers to be able
to hit the ground running with thermostat Eclipse development.

Note that in order for this to work the following Fedora bugs need to be
resolved. I can provide scratch builds if you like.

1. org.hamcrest: https://bugzilla.redhat.com/show_bug.cgi?id=852068
2. net.sf.cglib: https://bugzilla.redhat.com/show_bug.cgi?id=854249
3. org.apche.lucene: https://bugzilla.redhat.com/show_bug.cgi?id=849138

1 and 2 are for running plug-in tests. 3. is required for the
com.redhat.thermostat.common.core bundle to resolve properly (IIRC).

Anyhow it's nice to have this sort of thing, since it makes it easier for users
to set up a development environment for thermostat Eclipse plug-in development.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003076.html