Fix wording in empty thread pie-chart tab.
I was playing around with the new pie chart feature and it said I should select
a thread in the thread table. As I did so (clicked one thread in the table,
switched back to chart tab), nothing changed. When I double clicked on an entry
the pie-chart appeared.

This patch fixes the wording which was a bit misleading. I.e. it now mentiones
double-clicking explicitly.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003100.html