Fix build.
The jacoco maven plug-in uses argLine as the default property to use and it was
overwritten by the maven surefire plugin's argLine config. This patch adds
config for jacoco so that it uses coverageAgent as the property name over
argLine which should fix the build.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003118.html