Start netty as bundle.
This patch removes packages exported by the netty bundle from
osgi-export.properties and starts it as a bundle. Netty is required for
common-command. Any bundle depending on it needs netty started as bundle. In
order to be able to refer to things in command.properties file in a
version-less manner I've added a symlink to the actual jar.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/003797.html