Handle invalid command names
The launcher crashes with a NullPointerException if an invalid
command name is provided. This also happens in the shell. This
is a regression.

Fix it by throwing and handling an exception indicating the command
information is not available for that command.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/003832.html