Move and clean-up DbService
This patch moves DbService into the common-core bundle. It was in the launcher
bundle earlier. That service being in common-core reduces deps for the Eclipse
client.

I've also fixed the DbService interface so that no ServiceRegistrations are
accessible any longer. connect() registers the DbService and disconnect()
unregisters the service respectively.

Finally there was a small bug in StubBundleContext where it used equals() on a
String and an Object. Discouvered this during testing.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/003954.html