Fix NPE on Eclipse shutdown
Currently, every time the Eclipse plugins are stopped (ie. when Eclipse
is shutdown), an NPE is thrown while trying to disconnect from
DbService. This commit changes the catch clause for an unsuccessful
disconnect to use the Logger instead of printing to the console.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/004030.html