Pass Composite to Eclipse view constructors and remove synchronization
The current implementation for Eclipse Thermostat views passes a parent
Composite control to the view via a createControl method to be called
after the corresponding controller creates the view. This causes a
problem where the views methods are called by the controller before we
can call the createControl method. The solution implemented was to use
synchronization to wait for the createControl method to be called before
modifying dependent controls in the view.

This commit changes the above implementation to a much simpler approach.
This commit adds a SWTViewProvider interface for each view's ViewProvider
to implement. This interface provides a setParent method to set a
Composite to be used by the ViewProvider when instantiating the view
during createView. Now the views can create their controls in the
constructor, which allows us to remove the synchronization previously
required.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-November/004060.html