Fix test-deps repo missing error in Eclipse build
This commit fixes a build error introduced in [1], where the Eclipse
parent POM file lists the test dependencies p2 repository, but this repo
has not been built yet. In the changes introduced in [1], we build and
run the tests in a second step invoked after the Eclipse build. This
error only appears on a fresh clone when there is no existing p2
repository in the build tree, thus it slipped by before. This commit
moves the p2 repository specification from the Eclipse parent POM to the
POMs for the test plugins.

[1]
http://icedtea.classpath.org/pipermail/thermostat/2012-November/004063.html

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-November/004145.html