Move mongo storage bindings into separate bundle.
This patch moves MongoDB specific bits out of the common-core bundle. Some
config classes have been moved to storage-core in order to break a dep cycle.
Also, MongoConnection was catching NotImplementedException which is in
common-core so can't be used in the new bundle. However, I don't see how this
exception can occur in the first place so removing this unnecessary catch
seemed to be the right call. Note that existing tests continue to pass. Please
speak up if you think this NotImplementedException catch is necessary in
MongoConnection.

Mongo* classes are now all in an internal package.

With this it should be fairly easy to switch to an alternative backing storage
implementation as common-core is no longer mongo-java-driver dependent (at
least directly).

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-November/004292.html