Move VM Overview to its own bundle
This commit moves the VM Overview controller/views to its own bundles
(client-core and client-swing) under vm-overview in the top-level of the
Thermostat source tree. The client-swing bundle dynamically registers
the controller via a VmInformationService. This commit also moves the
SimpleTable and ChangeableText Swing components to this bundle, as these
classes are not publicly accessible and only used by this view. This
commit also adds a test for VmOverviewController as there was no test
class before.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-November/004284.html