Create VM Heap Analysis plugin
This commit consolidates the heapdumper functionality of Thermostat into
a self-contained plugin. Similar to previous commits, this moves the
heapdumper client code to a top-level directory. The commit also splits
the activator into a client-core activator which registers the
VmInformationService and the client-swing activator which registers the
various view providers.

There are a number of other changes in this plugin. The agent component
of the heapdumper is moved to an agent bundle in the plugin. The command
related functionality is split away from the client-core bundle into a
command bundle. The client-core code is then modified to not directly
invoke command classes, but obtain the command via an OSGi service. This
is done to not expose any heap command classes as API. Since there is a
little more setup required to do this in a generic way, I have created a
HeapDumper class. This class is responsible for setting up and running
the dump-heap command. I have also added a test for this class.

One other note is that since both command and client-core classes use
PrintObjectUtils, I have duplicated this class in both bundles. This is
done since this class is very small and it doesn't seem necessary to
make it API.

Reviewed-by: omajid, vanaltj, jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004389.html