Add TLS support to command channel.
This patch adds TLS support for the command channel. SSL for cmd-channel
communication is only used if things have been configured that way in
$THERMOSTAT_HOME/etc/ssl.properties. The reason as to why this has been done is
because if this is always on, users would need to set up a proper server (i.e.
agent) side keystore which is used for serving SSL enabled communication. An
alternative would be to ship with a default keystore. However, this defeats the
purpose of securing the channel in the first place. The world would have access
to this keystore. That's why we've opted for default == false. Note that these
defaults can be changed by packagers if they so desire.

If COMMAND_CHANNEL_USE_SSL=true then appropriate server and client SSL contexts
are created and passed to netty's SSLHandler. After a connection to the agent
was successfully established an SSL handshake is initiated (if so configured).

Tests have been added for most (all?) things I've added for SSL-enablement of
the command channel.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004770.html
PR1239