Don't start mongo HTTP status page.
If mongod is started on port 27518 say, then if the --nohttpinterface
option is not specified it also fires up a basic HTTP interface (similar
to httpd's server-status page) on port 28518. We don't need/want this
for thermostat storage.

This patch adds this option to common mongod options.

Reviewed-by: rkennke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004941.html