Create Ordered interface, sort Backends by ordering
This commit follows up on the addition of a ordering mechanism to sort
views in the Swing GUI. Since there are potentially many cases where we
want to order registered services when displaying information to the
user, I have moved the ordering mechanism to an interface that can be
used in each of these cases. The primary motivation for doing this is
the Backends table that is shown in the Agent information window.
Currently this only contains the System Backend, but subsequent commites
will split this up creating several Backends. This ordering mechanism
will allow us to display the Backends to the user in a predetermined order.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/004974.html