Create VM CPU agent and common bundles
This commit extracts VM CPU data collection from SystemBackend into a
vm-cpu-agent bundle. It also moves the VmCpuStatDAO from common-core
into a vm-cpu-common bundle that registers the DAO once Storage is
available. This also removes the DAO from DAOFactory.

One side-effect is changing the visibility of VmLatestPojoListGetter to
public in order for this DAO implementation (and DAOs in the other
commits in this series) to access it.

This also temporarily adds the vm-cpu-common bundle as a dependency to
client-cli until my later dynamic vm-stat patch.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/004996.html