Make InformationService impls internal
This commit moves the InformationService implementations used in each
Thermostat plug-in to internal packages.

Additionally, the commit modifies the respective Activators to track the
associated ViewProviders which are then passed to the constructor of the
InformationService implementation. This makes the dependency on the
ViewProvider(s) explicit and removes several uses of OSGIUtils.
Currently, an NPE is possible if the ViewProvider is not registered in time.

This commit requires several changes to the Eclipse client. Since each
Thermostat plug-in maps to an Eclipse view, the Eclipse view needs to
know which InformationService it is responsible for. This change to make
the InformationService implementations internal means Eclipse can no
longer reference specific implementation classes (e.g. the Host CPU
Eclipse view can't reference the HostCpuService). This commit adds an
identifier when the service is registered, so the Eclipse views can
query this identifier using an OSGi filter. This allows the Eclipse
views to still obtain the specific service instances they need without
referencing the implementation class.

These changes have also allowed more code to be pulled up into
RefViewPart so there is less duplication in the concrete classes.

Reviewed-by: vanaltj, jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/005296.html