Fix Swing Exceptions on connection failure
I recently noticed that when starting the GUI without the agent running,
sporadic exceptions would occur in Swing code. This seems to be caused
by interleaving of two threads calling ConnectionHandler.changed. When
DbService.connect fails to connect it throws a ConnectionException which
was caught by Main.connect. In this catch clause, we are calling
ConnectionHandler.changed, but at the same time the ConnectionListener
fires. This causes the handler to be called twice from different
threads. With this code removed, the sporadic exceptions no longer seem
to occur.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-February/005901.html