Convert Heap command.properties to plugin.xml
This commit moves the definitions for the heap-related commands to the
heap plugin's plugin.xml file. I have also updated the list-heap-dumps
usage string to reflect that hostId and vmId are optional arguments, and
fixed a spacing issue in the description for the logLevel built-in argument.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-February/005933.html