Automatically update Hosts/VMs tree in Eclipse
This commit adds a timer to the HostsVmsTreeViewPart in Eclipse, similar
to what the Swing GUI uses, to update the list of Hosts/VMs that are
shown to the user. This commit makes it possible to monitor new VMs that
are created while Eclipse is running. I have also changed the default
behaviour to only show live hosts, since this is what is done by default
in the Swing GUI. A mechanism to configure this option will come later.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-March/006197.html
PR1355