Clear THERMOSTAT_HOME in WebStorageEndpointTest
In WebStorageEndpointTest THERMOSTAT_HOME is set as a system property in
a setup method, but is never cleared during teardown. This results in
the variable still being set for other tests, and causes a test to fail
in Severin's JAAS patch if this test executed before it. This commit
simply clears the property in the teardown method.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-May/006529.html