Make SSL misconfiguration fatal
After having mistyped the password for my keystore used for Thermostat,
I noticed that these kinds of misconfigurations cause the agent to hang.
This is because after failing to create a TrustManager using information
from the ssl.properties file, we still attempt to connect. This commit
propagates the errors to the connection class, so that we can show the
user what is wrong with their configuration (wrong password, keystore
doesn't exist), and then exit.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-May/006580.html