Gracefully shutdown agent upon encountering an error
This commit moves the connection error handling in AgentApplication from
the ConnectionListener to a try/catch block. Currently, if the agent
fails to connect, the callback in ConnectionListener is triggered but
then a ConnectionException is thrown and we never catch it. When using
the service command, this ConnectionException prevents storage from
shutting down properly. I have also taken the opportunity to convert the
logger to a private static field, as is common elsewhere in Thermostat.
Also instead of passing the CountDownLatch to our SignalHandler, I
converted the SignalHandler to a non-static nested class and use the
latch field directly.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-May/006585.html