Convert VM IDs to UUIDs
This commit changes VM IDs used in VmRef and VmInfo to a randomly
generated UUID string. The VM PIDs are still stored, but are not used to
identify a specific VM. The motivation behind this change is that PIDs
can be reused by the OS, thus there is a chance we could have a PID that
refers to two different VMs (one alive, one dead). There are then
security implications if we return the wrong VM, which the client user
may not have access to. The VM's PID is still required when dealing with
external interfaces like /proc and JMX, but the data collected is placed
in storage using the VM's UUID.

Additionally, I have cleaned up the API in VmRef. I have removed
getIdString from VmRef since we already have a getStringID that does the
same thing. I've also changed VmRef.getAgent to getHostRef to more
accurately reflect what is returned.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-July/007242.html