Deserialization fix for PreparedParameters class
When running the webservice command, the new prepared statement
serialization fails while trying to deserialize a PreparedParameters
object. This is because there is no default constructor. This commit
simply adds a default constructor to be used only be the deserializer.

This problem doesn't appear when using the WAR, only the webservice
command. Maybe this is due to differences in class loading. Even if we
are not planning to support the webservice command, this commit shouldn't
cause harm and allows the webservice command to work properly.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-July/007636.html