Only send command channel request if authentication succeeds
While fixing the JMX plugin, I noticed that its RequestResponseListener
is receiving AUTH_FAILED responses twice. This is because even after the
generate-token fails with a given action, the request queue still issues
the command. This patch changes this behaviour so if generate-token
fails, no request is sent to the agent. This fixes the issue of
listeners being notified twice, and also has the benefit of not sending
requests that are destined to fail.

Reviewed-by: vanaltj, jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-July/007725.html