Convert RMIRegistry to an OSGi service
This patch converts the RMIRegistry class used by the agent proxy into
an OSGi service. This will allow all Thermostat agent bundles to easily
share the same RMI registry, which is restricted to localhost only. The
motivation for this patch is the future plan to add a proxy to
system-backend to allow a special "thermostat" user to receive
monitoring updates for other users' processes with jvmstat.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-November/008864.html