Separate storage command from agent-cli
With PR1608, we get an exception that the RMI registry port is already
in use for running "thermostat storage --status". This happens because
agent-core's activator creates a thread pool for JMX connections, which
in turn creates the RMI registry. My proposed solution is to separate
the storage command from the agent bundles, currently it resides in
agent-cli. This patch creates a storage-cli bundle that houses
StorageCommand and its related classes. There are then some
modifications to the ServiceCommand in order to invoke the storage
command without having a bundle dependency on storage-cli, which would
also entail having to export StorageCommand.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-December/009022.html
PR1608