Do not purge by default and warn user of potential clean-data impact
This is short term solution to PR2039.  The default agent configuration
is changed so that data is *not* purged on exit, as this can be
disruptive to other clients of the Storage node.  Furthermore, a warning
about this disruption is added to the clean-data command.

reviewed-by: jerboaa
review-thread: http://icedtea.classpath.org/pipermail/thermostat/2014-October/011337.html