Vm stat remove host option and add testNoVmId()
This is a patch that removes the unnecessary host option that the user is
prompted to enter for vmstat.  It also adds a a unit test to
VmStatCommandTest.java which checks for correct behaviour in the case where
no vmId argument is passed to the command.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-July/014607.html