Fix the error message when profiler can not find jars
The constructor `new JarFile()` can throw an IOException. When that happens,
the catch block raises an AssertionError. The message in the error uses the
uninitialized `jarFile` object, which results in a "null" in the message
string. The fix is simple: use the `jarPath` variable (just like it is already
used in the `System.out.println()` above, which is known to be defined and
initialized in the current scope.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-September/015956.html