Failing to launch the cmd-channel process should fail the agent
This patch reworks the command channel process in order for the agent to
shutdown when the command channel server fails to start. This is done by
the server process outputting a new '<SERVER STARTED>' token once the
command channel is ready to accept requests. Should the process fail to
start (e.g. missing classes on classpath), the agent will get an EOF
instead of this token and shut itself down.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-September/016178.html