Fix classpath for thermostat-command-channel in Fedora package
This patch fixes an issue in Fedora packaging where Netty is not found
on the classpath for thermostat-command-channel. This is due to
differences between Netty's version string in the Maven build and RPM
build. This fix adds a unversioned symlink to the Netty jar that the
command channel script can use.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-September/016464.html