[PATCH] fix profiling and heap dump on Windows
Both profiling and heap dump require the creation of files or directories within the user's temp directory (as defined by "java.io.tmp"(

Current Thermostat code uses PosixFilePermissions, which fails hard at runtime on Windows.

This patch removes the use of PosixFilePermissions on Windows only, allowing Heap Dump to work, and giving Profiling a better chance of working (more work is to be done there; I'm afraid that will be another patch)

On Windows, the default temp directory is private to the user, and secure.   Further patches will check if java.io.tmpdir is secure, and log a warning if it isn't, and probably allow the user to configure a global directory secured with an ACL if required to debug 'foreign' JVMs.

(This means that further work will need to be done to connect to another user's JVM for some Thermostat features that rely on the temp directory)

reviewed-by: sgehwolf
review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-February/022241.html